Finish setting up examples
