--- build.yaml
@@ -1,6 +1,0 @@
-targets:
-  $default:
-    sources:
-      include:
-        - lib/**
-        - widgetbook_testing_stuff/**
