+++ example/lib/main.dart
@@ -1,29 +1,9 @@
-import 'package:flutter/material.dart';
-import 'package:widgetbook/widgetbook.dart';
-import 'package:widgetbook_annotation/widgetbook_annotation.dart' as widgetbook;
-import 'package:example/l10n/app_localizations.dart';
+import 'package:example/widgetbook.directories.g.dart';
+import 'package:widgetbook_golden_test/widgetbook_golden_test.dart';
 
-import 'main.directories.g.dart';
-
-void main() async {
-  runApp(const WidgetbookApp());
-}
-
-@widgetbook.App()
-class WidgetbookApp extends StatelessWidget {
-  const WidgetbookApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Widgetbook.material(
-      directories: directories,
-      addons: [
-        LocalizationAddon(
-          locales: AppLocalizations.supportedLocales,
-          localizationsDelegates: AppLocalizations.localizationsDelegates,
-          initialLocale: AppLocalizations.supportedLocales.first,
-        ),
-      ],
-    );
-  }
+void main() {
+  runWidgetbookGoldenTests(
+    nodes: directories,
+    properties: WidgetbookGoldenTestsProperties(),
+  );
 }
