+++ example/lib/widgetbook_golden_test_example.dart
@@ -1,6 +1,6 @@
-import 'package:flutter/material.dart';
 import 'package:example/l10n/app_localizations.dart';
 import 'package:example/widgetbook.directories.g.dart';
+import 'package:flutter/material.dart';
 import 'package:widgetbook_golden_test/widgetbook_golden_test.dart';
 
 void main() {
