+++ example/pubspec.yaml
@@ -14,6 +14,8 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   intl: ^0.20.2
+  widgetbook_golden_test:
+    path: ../
 
 dev_dependencies:
   flutter_test:
@@ -21,8 +23,6 @@ dev_dependencies:
   flutter_lints: ^5.0.0
   widgetbook_generator: ^3.9.0
   build_runner: ^2.4.9
-  widgetbook_golden_test:
-    path: ../
 
 
 flutter:
