+++ test/integration/run_widgetbook_golden_tests_test.dart
@@ -1,4 +1,4 @@
-import 'package:example/main.directories.g.dart';
+import 'package:example/widgetbook.directories.g.dart';
 import 'package:widgetbook_golden_test/src/run_widgetbook_golden_tests.dart';
 import 'package:widgetbook_golden_test/src/widgetbook_golden_tests_properties.dart';
